requirements.txt
blinker==1.9.0
click==8.3.2
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
Werkzeug==3.1.8
WTForms==3.2.1
